Include the fivetran schema in the table name (#27802)
Fivetran has a notion of "schema" which is a way of grouping tables together. It is essentially part of the namespace of the table name (eg `default.teams` in postgres).

Since convex doesn't have this concept, map it on with an underscore.

Eg if schema is "fivetran_metadata" and table is "accounts", then the
convex table would be

Eg "fivetran_metadata_accounts"

![image](https://github.com/user-attachments/assets/cdfea9f1-3a30-4e70-a468-f53c5448b032)

## Test Plan

I manually tested with the mock destination connector
https://www.notion.so/convex-dev/Testing-the-Fivetran-destination-connector-892ba3c6a07f471391491740046995b2

![image](https://github.com/user-attachments/assets/c105bf45-8866-4f72-b411-0b2008534a46)

Fixes ENG-6916

GitOrigin-RevId: b8cd34561298b9d5e7505cac3f8f71a3ab6ec1ee