Fix URL validation (#27905)
Relax the check that URL end with ".convex.cloud" to allow for custom domains. Make sure the URL is valid URL and doesn't end up with convex.site.

Reuse validation code between next.js and common. This tweaks the error message slightly but I think this is reasonable to avoid redundant code.

GitOrigin-RevId: ab7dad10118ef97b8f9e79f4e0931f2c0833c285