allow virtual table and system table to share table numbers (#28008)
lift the requirement that virtual and physical tables have different table numbers, in one case: a virtual table and its corresponding physical table are allowed to have the same table number.

this does not affect table number assignment, so new virtual tables and system tables will still have distinct table numbers.

one potential concern is that `db.system.get` could see a virtual table id and think it's a physical system table id, because they have the same table number. This doesn't end up being a problem though because we always check if an ID is virtual before we try to look it up in the physical TableMapping.

tested as part of #27538 which includes the migration to make the table numbers match.

GitOrigin-RevId: 1558ece1048a9ea19abd1d87fdbfafc7fd424a3b