delete source_package/tests.rs (#28141)
this test has not run for over a year, since there is no `mod tests;` in the mod.rs.
It doesn't compile. I tried to get it to compile but eventually realized that it's testing behavior we don't really have anymore.
The behavior we want to test is covered in `application/src/tests/source_package.rs:test_source_package`

GitOrigin-RevId: 4027da6353716a8708f69d2bb7d1686c54e1d926