[transaction] rollback pending writes on child mutation failure (#28085)
wrap Transaction's data structures that hold pending writes in a `NestedWrites` struct, that keeps the current pending writes and a snapshot of the parent transaction's. So we can create child transactions, and put all new pending writes in there, and choose to either commit or roll them back.

With copy-on-write data structures the cloning to create the snapshot should be reasonably efficient. And it only matters if you're calling subtransactions anyway so I'm not worried about performance.

A larger potential issue is wrapping the wrong parts of Transaction. I believe I wrapped only things that are either readonly (like DatabaseSnapshot) and therefore don't matter if they are rolled back, or things that include pending writes. Notably I did not wrap the TransactionReadSet, which should get updated even if the nested mutation rolls back.

see new integration test. also tested with `just run-backend`

GitOrigin-RevId: 724dd6141628f1456715cf3d39c78030b4b2f871