Fix fivetran serialization issue (#28168)
There are two serde Serializer implementations for ConvexObject. One is
for the wire, vs the other for system metadata. We were using the wrong
one. Added a roundtrip test to help catch the bug.

I also manually tested with the MockFivetran connector to make sure it reproes / fixes original issue.

GitOrigin-RevId: a562a001bebeea4523f49bfb78ed2474635731ff