set default physical table numbers to virtual (#28138)
when creating new physical tables that have virtual tables, give them the same table number.

this is covered by a bunch of tests :) since it affects how new databases are created. also tested with local backend and i was able to upload a file to storage and access it with `ctx.db.system.query("_storage").collect()`

GitOrigin-RevId: 767c068816e3abc8bce828d1f45af4d55a83ab84