Move push logic to application layer (#28231)
This PR moves push logic from the backend layer to the application layer so that we can use it in setting up application tests with components. I also took the opportunity to break up the logic into helper functions. Analyzing modules requires the local node executor which depends on tokio, so the tests will need to use `ProdRuntime`.

GitOrigin-RevId: 618dbfda1ee5410aa6795e286564292a057f60dc