`npx convex codegen` with components (#28383)
update `npx convex codegen` to work in component-aware projects. It does this by running the first part of a component-aware push, including the `/start_push` api call to run analyze. This code is shared between `npx convex dev`, `npx convex deploy`, and now `npx convex codegen`.

one major change this implies: `npx convex codegen` must talk to a deployment now. this deployment is picked in the same way as `npx convex dev` picks it. in particular, smoke-tests and such need to start up a deployment and pass the `--url` and `--admin-key` flags

GitOrigin-RevId: 9e0ba4f8ba468e24bb38782cb0c7ca121b8394cd