[Expose Function Metadata] Working prototype for function metadata CLI command (#28235)
This is a rough prototype, I am very open to any suggestions.

Adds an API that can be used to expose function metadata. This is exposed in the form of a system udf. You can access the results through the CLI (`npx convex api-spec`) .

Two assumptions made in this PR are that:
- `any` is the default args validator value
- `any` is the default return validator value

We may decide that we only want to return functions that have both args and returns validators.

GitOrigin-RevId: 381916295dc8361f19bb36397da7dc5935833a66