fix auth.config.ts for components (#28479)
add auth.config.ts to modules in the `/deploy2` apis in the same way it's included to the `/push_config` api.

this fixes auth with components, because the CLI was never sending up the auth config bundle at all, and with this PR it does.

I would like to do the same pattern for schema.ts/schema.js but that can come later.

GitOrigin-RevId: a4875348bfa7903b64ee68bb43b956adae593d74