[Expose Function Metadata] Add new endpoint (#28463)
Adds endpoint that allows you to make a request to `api/run/messages/list` instead of including `messages:list` in the body as an argument. This endpoint can be used to call any public functions.

If an admin key is passed in, it can be used to call internal functions. This mimics the behavior of `api/query`, `api/mutation`, and `api/action`.

Also added relevant tests.

GitOrigin-RevId: 90a5b58100833e7694da18db0cf4e330c4152940