[virtual table cleanup] avoid reading from VirtualTableMapping (#28838)
VirtualTableMapping is redundant now; it can be derived from TableMapping and VirtualSystemMapping. Stop reading from it so we can stop writing to it.

The only remaining place we read from it is to support idempotent inserts, which only happen in backend. This should be safe to remove later without risk of version skew.

GitOrigin-RevId: 7fc33a36b4caccfac7f7dd5ac8b75d5294c41a7c