only fingerprint query when we're paginating (#28837)
i tracked down hot CPU in queries to the query fingerprinting, which does serialization and a sha256 hash. we should avoid computing the fingerprint when we don't need it, which is most of the time.

While I'm here, I also cleaned up `should_compute_split_cursor` by making a new enum which ties what the query does to how it will be used. If you're not paginating, you don't need fingerprint or any cursors. If you're manually paginating, you need fingerprint and cursor but not `end_cursor`. and if you're paginating reactively, you need fingerprint and `end_cursor` and split cursor.

GitOrigin-RevId: ff17b464763131c33abe5c69ad8e8285539e3e5a