running functions use ComponentId, not path (#28872)
When a function is ready to run, it should run with a ComponentId, not a ComponentPath.
ComponentPath is a way of referring to a mounted component, and sometimes we want to run functions in unmounted components. In particular there's now a way for system UDFs to request running such a function: `PublicFunctionPath::ResolvedComponent`.

in addition to existing tests, I manually tested the dashboard and made sure system UDFs components still work.

GitOrigin-RevId: 48074ae3ea3b71dd9be74a8c7f92843735ce1da6