Properly propagate Sentry context to the sync worker (#28961)
I think this is the correct way to ensure that Sentry events sent from the sync worker properly contain the context around client_id etc.
The way we were doing it before is a bit of a hack, but I'm leaving that for now too just in case this doesn't actually help.

GitOrigin-RevId: 587c1f30278d0c11c240493cc3f5e2bd9ca0b28d