delete VirtualTableMapping (#28665)
Now that all virtual tables have the same table numbers as their physical tables, and we have stopped reading from the virtual table mapping, it's time to delete it!

this PR has the effect of stopping writing to `_virtual_tables`. That should be fine because nothing should be reading that table (besides db-verifier and db-info and I can clean those up later).

We are currently reading from `_virtual_tables` but only for the purpose of making the insertions on backend initialization idempotent. So i think it's fine to delete all of this together -- it should be rollback safe.

GitOrigin-RevId: f7fd1ff015060df35bd3d68a66d84ffe23d604d8