npx convex codegen require a deployment (#28841)
in component-aware projects, `npx convex codegen` needs a deployment. currently we create a dev deployment if it doesn't exist, but this goes through the flow that is specific to `npx convex dev`. instead, we should go through the flow that all other commands go through that need a deployment, like `npx convex import`, `npx convex run`, and `npx convex deploy`.

GitOrigin-RevId: 271845bfea1be218c06300a6ec1b0b30686255a3