Log content type to event streams (#28999)
Gets ContentType and StorageId in the logging information.
Ideally it would include `Id<_storage>`, but it's a bit annoying to
get, so just doing StorageUuid for now which is 1-1 equivalent
for the time being.

This makes changes to both usher/backend - which I ensure compatibility of via the `optional` field.

GitOrigin-RevId: a9d40d4a790348c25850d22893b8cbe9b2361348