[Day.js Support] Support more of `Date` in our default runtime (#28992)
Pins all of our binaries to UTC via the `TZ` environment variable, so we can ensure determinism within all calls to the `Date` function. This will allow us to support all `Date` functions within our runtime.

This is a big change, so I want to make sure we are comfortable doing this.

GitOrigin-RevId: 52847f1bb848f08da2182ce4132dd93ec7575eb8