[ENG-7117] rethrow ConvexError from components (#29209)
this may not be how we want to handle error propagation eventually, so i left the TODO in. But this is an improvement on the current state, which doesn't propagate ConvexError or stacktraces at all. The new tests indicate that the stacktraces and ConvexError data are propagated the same as when an action calls `runQuery` (within the same component) today.

GitOrigin-RevId: 2d2ff53ec4236934d3601baeaaa6e95a64cdf665