test codegen on components projects (#29433)
run codegen wherever we see a `convex/_generated` folder, which indicates an app that can be codegened.

in particular, since we'll be iterating a lot on these and should make sure the codegen is correct, we should run codegen on the `components-poc` project and all example apps in the `npm-packages/components` directory. Since the example apps will be using the components directly instead of their packaged forms, we can use `just regenerate-codegen` to keep the `component/_generated` folders up-to-date.

to enable the test, fix a small issue from #29394 where we started running codegen on packaged components.

GitOrigin-RevId: d5f18125c69f551013b83a7d623e140872edf3ff