Baseline ComponentRegistry, still unused (#29380)
copy the pattern of SchemaRegistry to keep a ComponentRegistry which is updated with the latest state.
It's not used for any queries yet (not that Components are used in prod at all yet anyway) because I'm still trying to think about how we can generalize the query pipeline. Even the stuff in this PR should be generalizable.

- There should be a Struct of all registries that the Snapshot and Transaction can store
- There should be a way to do queries against the registry when you have a Snapshot. These will not take read dependencies.
- There should be a way to do queries against the registry when you have a Transaction. these *will* take read dependencies.
- We should have a way to validate that read deps are taken on the correct ranges. Even constructing the read dep for a simple lookup-by-ComponentId query was kinda tricky.

Anyway those are things I'm thinking about. I think this PR should be ready for review and we can start adding queries in later PRs when we discover the queries that are expensive or need to be done when we don't have a Transaction or read set.

GitOrigin-RevId: 552f77dff79a630ee13d11a4642708fa38c716e2