Use the convex codegen serialization for exports (#29444)
Use the convenient serialization/deserialization code.
Manually tested doing an export on `main` and viewing it from here. Worked great.

Don't love the `snake_case` - but that's what it takes to be backward compatible
here. Perhaps we can deserialize either and serialize camelCase and run a migration
to go to our more favored format, but figure I'll do that later.

GitOrigin-RevId: faee12ef44b26022b961bc6ba4294e9591bc6621