Exclude IndexDiff from component deployment audit log events (#29508)
This PR removes `IndexDiff` from `ComponentDiff` and from the dashboard displaying component push audit log events. `IndexDiff` is redundant because we already have the schema diff, and it's hard to keep track of because of 2 phase commit during push where indexes are added during `start_push` and dropped in `finish_push`.

GitOrigin-RevId: 220fa3a434f68ff8a338becf403ee33f0c4be984