snapshot import into a component (#29542)
pass a `--component-path` flag to `npx convex import` to import into the specified component.

I think component path instead of ID makes sense here because (1) it's writing data, which should generally only happen in mounted components, (2) it's user-visible input, (3) the rest of the import uses table names and component paths in the zip file's paths, so this is consistent.

added a test of importing a CSV into a child component. And manually checked that you can export a single component in the dashboard as a ZIP file, and import it back in-place.

GitOrigin-RevId: cbe874ecf58b921befafca9dbb6f47d42a04a262