snapshot import into missing component (#29567)
move `component_path_to_ids` to ComponentRegistry so it doesn't have to be async. Also it can return an `Option`, and `must_component_path_to_ids` is the new method that throws an error if the component doesn't exist.

Using this change, we can decide to handle missing component paths differently. In particular snapshot imports into missing components can throw a nice error.

Added a test.

GitOrigin-RevId: 3cdaa543cac01167b006586f672b8bab688db08a