clean up orphaned table namespaces (#29680)
In components push, we create a bunch of system and user tables in `start_push` that might not end up being used if `finish_push` is never called because of a validation error or other interruption. We should clean them up to avoid having tons of unused tables lying around.

GitOrigin-RevId: e73a29b9b435d70ae44131383445c18341a92700