add component id to action callback token (#29793)
This PR adds `ComponentId` to `ActionCallbacksToken` so that we can verify that node actions are running in a valid component. In a follow up, I'll hook in function references that will resolve relative to this `ComponentId` so we can call component functions from node actions in the root. Developers still won't be able to run node actions in component, so in practice this `ComponentId` we're passing through is always root for now.

GitOrigin-RevId: d1666912122d09a826cc55a2b25b87d17e889fa0