[ENG-7397] fix `Index on _schemas for {} not found` (#29983)
there's a race condition during `start_push` where a namespace and associated `_schemas` table is created before we load the `_schemas` table into memory. If you try to run a UDF during that time, it hits a system error.

I repro'd this locally by adding a 1 minute sleep in the right place and seeing the error.

To fix, we can say that `SchemaRegistry` just contains `_schemas` tables for components that exist. This works because the component is created in `finish_push`, after `start_push` completes, which implies the new schemas table has been pulled into memory. And it's safe because user tables in namespaces for components that don't exist yet cannot be written to, so the schema isn't useful for anything in the function runner.

With the fix, the repro no longer happens.

GitOrigin-RevId: 93e7be58d9505964441b3c417d38feae9f2eb3f9