pass RepeatableTimestamp through vector/text index backfill (#30292)
currently we downgrade RepeatableTimestamp (which is a Timestamp with a proof that it is repeatable) to a plain Timestamp and then try to upgrade it again by waiting for the committer to sync the proof to persistence. We can pass through the RepeatableTimestamp and that should work.

In the case where we read a Timestamp out of the database, we know that it was created from a `tx.begin_timestamp()`, so it's repeatable and we can use the in-memory Database (don't need it to be synced to persistence) to prove it with `tx.begin_timestamp().prior_ts()`

GitOrigin-RevId: 55fde9a5c7652d6afbac73cf66dfb2082009fba1