Add configuration for how eagerly to refresh auth tokens (#30588)
I would like this to be configurable so some things like Next.js + Convex Auth can refresh more eagerly than the default.

I made it so that if the configured period is shorter than the token lifetime, we'll just refetch immediately, but we'll log a warning since repeatedly fetching immediately is probably a misconfiguration.

GitOrigin-RevId: 10a9efbeaac3f91ef75f42923ce126e8730eeb41