Disallow excess arguments in CLI (#30631)
Disallow excess argument in the CLI to prevent accidental misuse and make it possible to deprecate options in the future. I noticed this when I tried to use arguments with `npx convex dev --run "messages:list" '{"arg": 1}'` and they were ignored; I was using the command wrong, we should help out with that.

We'll want to call this out in upgrade notes.

GitOrigin-RevId: f6b18bdee1ffabf556172488b909714689b5aaf8