[ENG-6401] [components] attach path to console logs from sub-queries and sub-mutations (#30679)
when running a mutation or query as a subfunction, the logs should get accumulated with the parent function but they should have the subfunction's path attached. I considered a few ways of doing this ( #30593 and #30576 ) but found the one in this PR to work well:

enhance the `LogLine` enum so it has a recursive `LogLine::SubFunction` variant.

Some places, like node_executor and LogEvent, only deal with the non-recursive variant of `LogLine`, now called `LogLineStructured`, but most can have a full tree of log lines.

Sometimes we need to serialize the tree structure: in protobuf for funrun & usher, and in ConvexValue for `_session_requests`. Meanwhile sometimes we need to flatten the tree into an array: when we need to call `to_pretty_string()` for `RedactedLogLines`, and when we output `LogEvent`s.

GitOrigin-RevId: 7274b546885003b4be35072082d12773e3c48c5c