Record a histogram for active DB connections instead of a gauge (#30837)
I keep seeing traces where it takes a while to acquire a connection, so am wondering how close we get to the connection limit. The current gauge seems to be somewhat misleading since it's only reported periodically. Another potential alternative is logging a counter whenever we're at the max connections or nearing the max connections.

GitOrigin-RevId: 2a77a5f85066d5a4a54f668b952842f53f0699e1