Add a JSON variation of tracing sampling config (#30924)
The string format is IMO confusing now that there are both instance overrides and route overrides.

The tests have a bunch of examples of what the new format looks like + its behavior.

A couple interesting notes:
* Named everything `fraction` + added errors if the value is ever something that's not between 0 and 1
* `defaultFraction` is technically redundant with a blanket `routeOverrides`, but it seems nice to be able to tell the default sample rate at a glance.
* `instanceOverrides` is kind of dumb when setting this value for backend, but it's optional and seems not worth the work to structure differently

I don't know if we have strict limits on the length of the knob value, but I'm hoping going for clarity over compactness is less error prone.

Both formats will work for now (using starts with `{` as the criteria for the JSON format), and once it's rolled out without chance for rollback, I'll change any knobs + the default value across our relevant services and delete the old code path.

GitOrigin-RevId: 607991e228a2b4b98cb36e95f8cc2141db831dee