Make maxDate optional in paginatedDeploymentEvents (#30950)
I want to update the logs page of the dashboard to use `paginatedDeploymentEvents` instead of not paginating so we don't hit the bytes read limit. To do this, I need to make maxDate optional because we always need the latest deployment events.

Once we update the dashboard to use this API, we can get rid of listDeploymentEventsFromTime

GitOrigin-RevId: b583a4227ed7cf10f7a1546abd721c74f6a78b9c