fix race in crons tests (#31020)
udf-tests contains crons, so the crons tests have a race:

- load the modules with crons
- crons start running for the first time in the background, reading and writing the `_crons` table
- then we create a new cron for the test and that can hit an OCC

this is currently not causing a failure, but when i change some internal tokio scheduling within queries, it starts failing. so let's preemptively fix it.

note this is using TestRuntime, so time jumps forward as much as it needs to, and the 100 second sleep takes a few milliseconds while running all crons that are scheduled to run immediately

GitOrigin-RevId: a433f7edb42c47719f924ea75902c912d0769fb8