lazily parse scheduled job args (#31053)
currently we parse udf args from Bytes -> ConvexArray whenever we read from the `_scheduled_jobs` table.
This can be unnecessary, for example `SchedulerModel::complete` reads the job, parses it, changes the state, and replaces the document. In that flow we don't need to parse and re-serialize udf arguments.

Only parse arguments when necessary: when returning a virtual table document, when running the function, or when reporting a system error to the logs.

GitOrigin-RevId: bdfadf8ea2adaa9e1ec3d42fbcb60397314f9aed