Increase MySQL connection pool TTLs for funrun (#31367)
We've seen a ~40ms overhead when acquiring a new connection (as opposed to reusing an existing one). The current parameters try and keep the pool small but are from an era before funrun as well as before we were using RDS proxy. This makes it so that funrun keeps connections around for longer so we can benefit from connection reuse.

GitOrigin-RevId: db9e30bcb89ab159bacfaccadf545d6b6547b9df