delay between runs of ScheduledJobGarbageCollector (#31511)
if scheduled jobs run frequently, after two weeks there is work for ScheduledJobGarbageCollector. If it finds garbage eligible for deletion, even if it's just a single job, it immediately reruns. And it doesn't reset backoff in this case.

We can reduce committer activity by debouncing. If we did some deletes, wait 10 seconds before rerunning, to allow more jobs to become eligible. And reset backoff in case there were errors previously.

GitOrigin-RevId: aa5e4eadc8097eb8b9fd97327d9f9c7c94bb72c5