Create new replace_all functionality for snapshot import (#31181)
Added ReplaceAll mode - which deletes anything that is preexisting that isn't in the import. Completely clears the deployment and replaces it with whatever is in the backup.

Main change is in `snapshot_import.rs`. Prior to the loop which streams the import, we iterate through all the active user tables and either delete or clear them depending on the presence of a schema.

Also added deleted tables to the deployment audit log. Fixed UI for it.

Added a test for it. Confirmed the test fails on Replace but pases on
ReplaceAll. The test does a ReplaceAll that moves a doc from one table to another
and also exercises the truncation of a table that's in the schema.

GitOrigin-RevId: ee76c599335b4e31c948764d7c9cd8b0f41488c9