Hook up sentry reporting for local deployments (#31478)
While we're rolling out dev with local deployments and trying to identify bugs, it'll be helpful to be able to see any errors that the local backend hits (similar to being able to see errors hit by `npx convex dev` and similar from cloud deployments).

This change makes it so that `npx convex dev --local` reports errors from the local backend to one of our Sentry projects -- alpha testing instructions will be updated to make this clear. We'll make this reporting opt in or at least configurable before releasing more broadly.

This makes it so that local backend can generally be configured to report errors (I think we weren't calling `sentry::init` anywhere before), but only if there's a `SENTRY_DSN` env var provided.

GitOrigin-RevId: f08666117de00545e3ef755fb682fdf15964f065