Add component awareness to deployment audit log for restores/imports (#31913)
Requires component awareness in a bunch of the stack. Fixes a bug with
the table-number-conflict-check not being component aware either.

updated the format in the deployment audit log and manually tested. Will
follow up with a migration to move to new format followed by removing
dead code.

Fixes ENG-7871
Fixes ENG-7872

GitOrigin-RevId: 3355712de22cd287a08adb449366c588fa77f374