Make report_error async. Yield once before reporting. (#31853)
Yielding allows the task to die during shutdown - which can
avoid spuriously reporting useless abort errors to sentry.
Should hopefully reduce a lot of noise during pushes.

I considered making it the responsibility of the callsite to yield, but given that every background worker would prefer this yield, and it doesn't hurt to add it on main threads - I just achieved it everywhere with this technique. Open to better ideas, but this seems to compile.

GitOrigin-RevId: 168f9e54d7f66e942959d51c0d17749820b03596