Add top k metrics for cache hit rate and failure rate (#32432)
Adds top k metrics for failure rate and cache hit rate to be used by the dashboard.

The result of these endpoints is a map of function identifier to timeseries, where there are `k+1` items. The `_rest` key is a timeseries of the sum of the metric for all functions not included in `k`

GitOrigin-RevId: eb20733e6d647ecb5821e9e8f47b3532b9c023a1