Add support for multiple onTransition handlers on BaseConvexClient (#32367)
This change is meant to accommodate some of the local store / object sync engine work.

The goal is to have a separate client library that uses the same `BaseConvexClient` to make requests to the Convex backend, while also allowing other Convex traffic via `ConvexReactClient` (or similar), but making it so that we externalize transitions at the same time for both clients.

We were initially going to split `onTransition` into an asynchronous `prepareTransition` phase and a synchronous `completeTransition`, where the latter externalizes the results (see https://github.com/get-convex/convex/pull/32308). Instead, we're going to keep this synchronous, but allow for multiple `onTransition` handlers and pass them slightly more information.

The main change here is passing around more information to the `onTransition` handlers, like the query updates vs. the just the tokens of the changed queries. The most interesting one is returning reflected mutations + the value returned from the mutation -- previously we closed over the result in a promise resolver, but now we need to track the result separately so we can feed it into these handlers.

GitOrigin-RevId: 4af4b061f0a8e4f545e3ca6419e72106b85a65a9