[1/?] Squash down the per-component export (#32762)
Split tables, storage, and tables table into sep funcs.

Change the order in which we write to the zip.
- Write all user tables (for all components)
- Write all tables tables (for all components)
- Write all storage tables (for all components)

This will allow us to write tables in any order (not necessarily in
component order) which may be useful for our optimizations. Next step
would be to sort all user tables by size across components (coming soon)

GitOrigin-RevId: 0bb1c5f2cf66f75b0c89bade29d0ef6a56f94131