reclassify repeatable_ts errors coming from funrun HttpActions (#32724)
we're seeing intermittent stacktraces indicating that funrun is receiving requests for HttpActions at timestamps that are older than min_snapshot_ts. A snapshot_ts being outside of retention is a known failure case which is classified as Overloaded.

But accidentally we're throwing the non-repeatable timestamp violation, which is scarier and doesn't apply. It applies when you request a read at a timestamp that is > the repeatable_ts. In this case the requested read is at the repeatable_ts, and we're just checking that min_snapshot_ts is below it.

GitOrigin-RevId: 402421d22c7a12b78f7e818a5d4ebdeb27d51b9a