[ENG-8227] fix test_new_sync_query_after_disconnect (#33104)
test was failing nondeterministically because `t.server.latest_timestamp().await?` is the timestamp of the latest mutation, and there was a chance for the mutation to propagate to the client before the `.add_sync_query("getConversations", assert_obj!())` fully executes. So there's not always a result.

The fix is to do the second mutation after the query is registered, so waiting for the client to be aware of the mutation also means the query must be ready.

GitOrigin-RevId: 9406d69ea542af5c27ac5ebf5846789b3489978e