Fix tracing bug and add write size event (#33197)
This PR fixes a bug in commit tracing where we would increment the `commit_id` even if we didn't push a persistence write to the queue. This fixed by making `start_commit` return an `Option<BoxFuture<..>>` to write to persistence and increment the `commit_id` if it is some. Also adds the size and number of documents written to the traces.

GitOrigin-RevId: d7eda2d624bfe005883d91ce60378f4f58e53466