avoid allocations in `WriteLog::iter` (#33251)
undo the part of #31531 which adds a lot of allocations to `WriteLog::iter`.
The idea from that PR is that `Vector` is copy-on-write, so it's cheap to clone within a mutex and then we can do expensive operations outside the mutex without worrying about races.

Unfortunately, `WriteLog::iter` was changed to pretty much do an element-wise clone of the whole write log. This shows up in polarsignals heap traces, and I expect it's not helping with performance.

Specifically, `Vector::slice` is a "write" operation for the purposes of copy-on-write.
We always call `WriteLog::iter().iter()` immediately anyway, so the lifetimes aren't an issue and we can go back to having `WirteLog::iter()` return an Iterator with references.

GitOrigin-RevId: b2e84309c106075f256cacc1ed087f38907b55f9