[query cache] fix race where unauthed query re-executes multiple times (#33298)
there's a race condition which is easy to hit now:

1. there's a query that doesn't read `ctx.auth`, and is requested by two clients 1 and 2 with different auth.
2. the data changes, so the two queries are invalidated together.
3. query1 checks the cache, finds the Ready cache entry is invalid, and removes the cache entry
4. query1 creates a Waiting entry for its specific auth, and starts executing
5. query2 checks the cache, finds no cache entry at all, and starts executing

so both queries execute, even though one of them should wait for the other.

To fix, we change step 4 to create a Waiting entry for the same cache key that it found to be invalid. That way, the second query will check the cache and wait for execution to finish. This doesn't affect the eventual cache key that gets saved; if query1 ends up checking auth, that will be stored in the cache and query2 will detect that the thing it's waiting for went away (same as if the thing it's waiting for hit an error) and re-execute.

This doesn't entirely fix the race condition, since steps 4 and 5 can have their order swapped and then we're back to re-executing. But it does make the race condition much tighter. If we actually wanted to fix it, we could consider changing `Ready` cache entries into `Invalidated` in step 3 instead of removing them completely.

Updated the test that covers this race condition, so now it asserts that there is a cache hit.

GitOrigin-RevId: 447d41c20d7586a527ad8dd6779b09d96300a339