Thread prev_ts through DocumentUpdate (#33295)
This adds prev_ts to the old_document of `tx.apply_validated_write`, requiring the caller to thread the ts from whatever document they queried. This is then threaded all the way through DocumentUpdate and into the persistence layer.

In particular, the caller passes a `WriteTimestamp`, which may be `Pending`; `Writes::update` is responsible for collapsing the prev_ts chain in that case.

Some users of DocumentUpdate - primarily around index updates - do not care about the prev_ts. I have split those into a separate type, `IndexDocumentUpdate`.

There is a push safety consideration because we serialize the write set and send it around. As such I have marked the prev_ts as optional almost everywhere to account for the case where it gets lost during an RPC. This can get cleaned up after a push cycle.

GitOrigin-RevId: f6cca7c59dce42d159780486d30142ebdae7d3c3