[import] fix bug when file storage table changes number (#33703)
follow up to #31156 -- allow the `_file_storage` table number to change. We already mostly supported this, but unfortunately when committing the file document we tried to look up table number -> tablet id. we never used the tablet id for anything, but if the table number is changing, this throws an error.

also we weren't passing through `table_mapping_for_schema` to the insert, which turns out fine because `_file_storage` doesn't have a schema, but we might as well be correct about it in case it eventually gets one.

GitOrigin-RevId: f66cd916a9814d6c6f9e809c7aba59a467269a30