dashboard: avoid convex analytics in self hosted dashboard product (#33710)
The convex dashboard occasionally logs product-focused analytics events when users perform actions, such as clicking on the "Run Function" button.

We use these analytics to track the usefulness of features so we can improve them.

This PR refactors the code in `dashboard-common` such that events are only logged in the self-hosted dashboard via dependency injection

We may re-add analytics in the dashboard self-hosted deployments in the future, but it would be an opt-in feature.

GitOrigin-RevId: 465cabc0b929b96c4ac827c105f0f3f25378f9cf