[components] fix bug when scheduling across components with different table numbers (#33780)
if you do a `ctx.scheduler.runAfter` from a scheduled function, to run a function in a different component, and the two components have different table numbers for `_scheduled_jobs`, it throws a system error.

That's because scheduled functions with parents check to see if their parent was canceled. but if the parent was in a different component, we don't know how to look up the parent job because we don't have the parent's component id.

This PR threads parent component ID through everywhere we have parent job id -- with defaults to ComponentId::Root for backwards compatibility -- and starts using it when looking up the parent job.

GitOrigin-RevId: bcd1e4d35de74f77140ffd99b7afc61aaf716cda