Rewrite self hosted backend Dockerfile (#33804)
Okay, this should be a bit closer to a "best-practice" Dockerfile!

The total compressed image size is ~160MB which seems very acceptable, though I added an optional build arg `debug` if you want to preserve symbols which makes it significantly larger.

It's reasonably quick to rebuild (which isn't important for our current goals, but certainly is nice to have) on Linux, though on Mac we're still limited by how quickly Docker can scan your filesystem for local changes (surprise: not quickly). This is, however, the iteration speed I'm used to with other Docker-based projects.

Tested that the image at least seems to work with `--help` -- I didn't try passing an admin key in yet.

During the build, it'll currently print out these lines:
```
#31 0.938 warning: local_backend@0.1.0: not within a suitable 'git' worktree!
#31 0.938 warning: local_backend@0.1.0: VERGEN_GIT_COMMIT_TIMESTAMP set to default
#31 0.938 warning: local_backend@0.1.0: VERGEN_GIT_SHA set to default
```
Because this is no longer a git repo inside the container, vergen is complaining that these were set to default. We can configure vergen not to warn and then in "real" builds, we'll just need to pass `VERGEN_GIT_COMMIT_TIMESTAMP`  and `VERGEN_GIT_SHA` ourselves to the `docker build` command (cc @nipunn1313)

GitOrigin-RevId: 2b743981dfc9586da34d199fb4ca794a84307b0c