[self-host] auto-generate instance secret (#33847)
before the backend starts, generate an instance secret and name if one is not provided or stored. Store the credentials in the data volume so they will be reused when the container restarts.

random instance secrets can be generated by `openssl rand -hex 32` and default instance name is picked to be "convex-self-hosted"

GitOrigin-RevId: a0f8b0fb7fb7af5122f89011a903e3245fd93d0f