npx convex self-host import|export (#34094)
just as with `npx convex self-host run`, refactor the command line options to command.ts, the deployment-agnostic behavior to lib/, and create new subcommands `self-host import` and `self-host export`.

one wrinkle is that we don't know the dashboard url from the self-hosted commands, so i adjusted the console output in the cases where we don't have a url to print out for cases like "import already in progress. see it at <dashboard link>"

GitOrigin-RevId: 15941e107aa9ade1bf15e84cccb56859260dc39b