[self-host] skip picking instance name (#34095)
the first paragraph says to pick an instance name, but then it's never used. instead, the instructions always cause the default instance name "convex-self-hosted" to be used.

This is usually fine. The instance name doesn't show up many places and you can't run multiple instances on a single backend (or even with the same docker data volume) so there's no harm in conflicts.

We can add a section later for advanced customization, which would include how to set instance name (through an env variable) and also how to extract the instance_secret if you need it for some reason.

GitOrigin-RevId: 48e865d627e0b49e4f0921971b4d55ab6d80d484