NullAccessTokenAuth throw unauthenticated error (#34225)
when trying to use an access token or invalid admin key against an OSS (non-big-brain-affiliated) backend, throw an authenticated error instead of an internal server error.

This is still valid because, while it's true that access tokens aren't checked by OSS instances (which could be called an internal server error), it's also true that they can't be created so any attempt to use one would be an authentication failure.

part of ENG-8418

GitOrigin-RevId: 52d2de9d96d44943d253adbe1358d08b3222c879