validate `previous_revisions` timestamp is within document retention (#34238)
currently we're validating that the timestamp is within document retention, but only for timestamps that actually have a previous revision. Consider if you query at a too-old-timestamp, where it's very likely the previous revision has been deleted by document retention. Then we don't assert anything, but this is exactly the case where we want to throw an error.

fix it by calculating the min_ts of all the input timestamps, not of all the ones that actually had results.

GitOrigin-RevId: 562ee684de507dc13bcec95e20670cb116ff8d65