use documents_multiget in retention (#34317)
Make the assumption that prev_ts has been written for all documents being looked at by index retention. Since index retention only lags by a few minutes (we have an alert if it lags by over an hour), all of these revisions must have been written with correct `prev_rev` fields.

That means instead of calling `previous_revisions` to get the previous revision for each document, which does a range scan on a secondary index, we can do `documents_multiget` which does point-lookups on the primary index.

Not only should this be inherently faster, it has the benefit of being a no-op if all of the log entries are inserts. Overall, this should make retention faster.

GitOrigin-RevId: b88ba60c6b1cde47312c376c4982a323fbab87d4