add `is_canonical` field to vanity_domains in bb db (#34324)
For canonical domains, we will need a place to store whether a domain has been marked as canonical. This is a user-configurable setting, where they select at most one custom domain to use "canonically" for .convex.cloud and one for .convex.site. These canonical domains don't take effect unless they are also verified, but whether it's marked as canonical is an independent boolean that we should store in the db.

GitOrigin-RevId: 4b2f6c4e2a52fea21e631cf771e32a18e0969491