dashboard: add navbar to project settings page (#34414)
Adds a sticky navbar to the project settings page that has a faux-scrollbar which scrolls alongside the main content to indicate how far you are scrolled in the navbar. Also, the elements that are currently in view are highlighted.

Clicking on a nav item scrolls that section into view.

URL hash matching one of the section ids will scroll directly to that section, making docs links easier to use.

GitOrigin-RevId: 12fea75c1d8da03c1125ccc5f74e7c0abde400be